services/cloud-provider/requirements-dev.txt
pytest==8.3.4
pytest-cov==4.1.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
moto[all]==5.0.25
black==25.12.0
flake8==7.3.0
mypy==1.19.1
boto3-stubs[essential]==1.35.76
